requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
python-decouple>=3.8
Pillow>=10.0.0
django-filter>=23.0
reportlab>=4.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0

# Para desenvolvimento (opcional)
django-debug-toolbar>=4.0.0
